# IGNITE-19542 .NET: Add BinaryTupleIgniteTupleAdapter