--- a/modules/platforms/dotnet/Apache.Ignite.Benchmarks/Program.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Benchmarks/Program.cs
@@ -19,8 +19,9 @@ namespace Apache.Ignite.Benchmarks;
 
 using BenchmarkDotNet.Running;
 using Table;
+using Table.Serialization;
 
 internal static class Program
 {
-    private static void Main() => BenchmarkRunner.Run<DataStreamerBenchmark>();
+    private static void Main() => BenchmarkRunner.Run<SerializerHandlerReadBenchmarks>();
 }