--- a/modules/platforms/dotnet/Apache.Ignite.Tests/Table/BinaryTupleIgniteTupleAdapterTests.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Tests/Table/BinaryTupleIgniteTupleAdapterTests.cs
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Apache.Ignite.Tests.Table;
+
+using System;
+using Ignite.Table;
+using Internal.Proto.BinaryTuple;
+using Internal.Table;
+
+/// <summary>
+/// Tests for <see cref="BinaryTupleIgniteTupleAdapterTests"/>. Ensures consistency with <see cref="IgniteTuple"/>.
+/// </summary>
+public class BinaryTupleIgniteTupleAdapterTests : IgniteTupleTests
+{
+    protected override IIgniteTuple CreateTuple(IIgniteTuple source)
+    {
+        // TODO: Serialize to BinaryTuple and wrap.
+        return new BinaryTupleIgniteTupleAdapter(Array.Empty<byte>(), new Schema(0, 0, 0, 0, Array.Empty<Column>()), 0);
+    }
+}