--- a/modules/platforms/dotnet/Apache.Ignite.Tests/Table/BinaryTupleIgniteTupleAdapterTests.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Tests/Table/BinaryTupleIgniteTupleAdapterTests.cs
@@ -31,6 +31,8 @@ using NUnit.Framework;
 [TestFixture]
 public class BinaryTupleIgniteTupleAdapterTests : IgniteTupleTests
 {
+    protected override string GetShortClassName() => nameof(BinaryTupleIgniteTupleAdapter);
+
     protected override IIgniteTuple CreateTuple(IIgniteTuple source)
     {
         var cols = new List<Column>();