--- a/modules/platforms/dotnet/Apache.Ignite.Tests/Table/IgniteTupleTests.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Tests/Table/IgniteTupleTests.cs
@@ -30,7 +30,7 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestCreateUpdateRead()
         {
-            IIgniteTuple tuple = new IgniteTuple();
+            IIgniteTuple tuple = CreateTuple(new IgniteTuple());
             Assert.AreEqual(0, tuple.FieldCount);
 
             tuple["foo"] = 1;
@@ -68,7 +68,7 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestGetNullOrEmptyNameThrowsException()
         {
-            var tuple = new IgniteTuple { ["Foo"] = 1 };
+            var tuple = CreateTuple(new IgniteTuple { ["Foo"] = 1 });
 
             var ex = Assert.Throws<ArgumentException>(() => tuple.GetOrdinal(string.Empty));
             Assert.AreEqual("Column name can not be null or empty.", ex!.Message);
@@ -92,7 +92,7 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestGetNonExistingNameThrowsException()
         {
-            var tuple = new IgniteTuple { ["Foo"] = 1 };
+            var tuple = CreateTuple(new IgniteTuple { ["Foo"] = 1 });
 
             var ex = Assert.Throws<KeyNotFoundException>(() =>
             {
@@ -104,24 +104,25 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestToStringEmpty()
         {
-            Assert.AreEqual("IgniteTuple { }", new IgniteTuple().ToString());
+            var tuple = CreateTuple(new IgniteTuple());
+            Assert.AreEqual("IgniteTuple { }", tuple.ToString());
         }
 
         [Test]
         public void TestToStringOneField()
         {
-            var tuple = new IgniteTuple { ["foo"] = 1 };
+            var tuple = CreateTuple(new IgniteTuple { ["foo"] = 1 });
             Assert.AreEqual("IgniteTuple { FOO = 1 }", tuple.ToString());
         }
 
         [Test]
         public void TestToStringTwoFields()
         {
-            var tuple = new IgniteTuple
+            var tuple = CreateTuple(new IgniteTuple
             {
                 ["foo"] = 1,
                 ["b"] = "abcd"
-            };
+            });
 
             Assert.AreEqual("IgniteTuple { FOO = 1, B = abcd }", tuple.ToString());
         }
@@ -129,9 +130,9 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestEquality()
         {
-            var t1 = new IgniteTuple(2) { ["k"] = 1, ["v"] = "2" };
-            var t2 = new IgniteTuple(3) { ["k"] = 1, ["v"] = "2" };
-            var t3 = new IgniteTuple(4) { ["k"] = 1, ["v"] = null };
+            var t1 = CreateTuple(new IgniteTuple(2) { ["k"] = 1, ["v"] = "2" });
+            var t2 = CreateTuple(new IgniteTuple(3) { ["k"] = 1, ["v"] = "2" });
+            var t3 = CreateTuple(new IgniteTuple(4) { ["k"] = 1, ["v"] = null });
 
             Assert.AreEqual(t1, t2);
             Assert.AreEqual(t2, t1);
@@ -147,11 +148,13 @@ namespace Apache.Ignite.Tests.Table
         [Test]
         public void TestCustomTupleEquality()
         {
-            var tuple = new IgniteTuple { ["key"] = 42L, ["val"] = "Val1" };
+            var tuple = CreateTuple(new IgniteTuple { ["key"] = 42L, ["val"] = "Val1" });
             var customTuple = new CustomTestIgniteTuple();
 
             Assert.IsTrue(IIgniteTuple.Equals(tuple, customTuple));
             Assert.AreEqual(IIgniteTuple.GetHashCode(tuple), IIgniteTuple.GetHashCode(customTuple));
         }
+
+        protected virtual IIgniteTuple CreateTuple(IIgniteTuple source) => source;
     }
 }