--- a/modules/platforms/dotnet/Apache.Ignite.Tests/ToStringTests.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Tests/ToStringTests.cs
@@ -52,7 +52,8 @@ public class ToStringTests
                 var code = File.ReadAllText(path);
 
                 if (code.Contains("new IgniteToStringBuilder(", StringComparison.Ordinal) ||
-                    code.Contains("IgniteToStringBuilder.Build(", StringComparison.Ordinal))
+                    code.Contains("IgniteToStringBuilder.Build(", StringComparison.Ordinal) ||
+                    code.Contains("IgniteTupleCommon.ToString(this)", StringComparison.Ordinal))
                 {
                     continue;
                 }