--- a/modules/platforms/dotnet/Apache.Ignite.Tests/ToStringTests.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Tests/ToStringTests.cs
@@ -53,7 +53,7 @@ public class ToStringTests
 
                 if (code.Contains("new IgniteToStringBuilder(", StringComparison.Ordinal) ||
                     code.Contains("IgniteToStringBuilder.Build(", StringComparison.Ordinal) ||
-                    code.Contains("IgniteTupleCommon.ToString(this)", StringComparison.Ordinal))
+                    code.Contains("IIgniteTuple.ToString(this)", StringComparison.Ordinal))
                 {
                     continue;
                 }