… modules/platforms/dotnet/Apache.Ignite/Internal/Table/IgniteTupleCommon.cs
… modules/platforms/dotnet/Apache.Ignite/Internal/Table/Schema.cs
… modules/platforms/dotnet/Apache.Ignite/Internal/Table/Serialization/TupleSerializerHandler.cs
… modules/platforms/dotnet/Apache.Ignite/Table/IIgniteTuple.cs
… modules/platforms/dotnet/Apache.Ignite/Table/IgniteTuple.cs