When reading data from server, currently we unpack all elements of the incoming `BinaryTuple` into `IgniteTuple`. This is extra work and extra allocations. Instead, create an adapter from `BinaryTuple` to `IIgniteTuple` like `MutableRowTupleAdapter` does in Java.

**Benchmarks**

```
BEFORE

|             Method |      Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Allocated |
|------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|
|   ReadObjectManual |  52.66 ns | 0.326 ns | 0.305 ns |  1.00 |    0.00 | 0.0003 |      80 B |
|         ReadObject |  88.41 ns | 0.425 ns | 0.398 ns |  1.68 |    0.01 | 0.0002 |      80 B |
|          ReadTuple | 263.26 ns | 2.822 ns | 2.502 ns |  5.00 |    0.06 | 0.0019 |     544 B |
| ReadTupleAndFields | 268.03 ns | 3.866 ns | 3.616 ns |  5.09 |    0.09 | 0.0019 |     544 B |


AFTER

|             Method |      Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Allocated |
|------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|
|   ReadObjectManual |  53.23 ns | 0.320 ns | 0.268 ns |  1.00 |    0.00 | 0.0003 |      80 B |
|         ReadObject |  88.01 ns | 0.484 ns | 0.453 ns |  1.65 |    0.01 | 0.0002 |      80 B |
|          ReadTuple |  23.66 ns | 0.274 ns | 0.257 ns |  0.44 |    0.00 | 0.0004 |     120 B |
| ReadTupleAndFields | 136.34 ns | 2.014 ns | 1.884 ns |  2.56 |    0.04 | 0.0007 |     208 B |
```

After this change, `ReadTuple` simply performs a buffer copy, but does not deserialize individual fields. `ReadTupleAndFields` unpacks all fields, and it is still faster and allocates less than before.